+++ crates/forge-runner/src/running.rs
@@ -46,7 +46,6 @@ use tokio::task::JoinHandle;
 use universal_sierra_compiler_api::AssembledProgramWithDebugInfo;
 
 pub mod config_run;
-mod copied_code;
 mod execution;
 mod hints;
 mod setup;
@@ -231,7 +230,6 @@ pub fn run_test_case(
         environment_variables: runtime_config.environment_variables,
         contracts_data: runtime_config.contracts_data,
         fuzzer_rng,
-        experimental_oracles_enabled: runtime_config.experimental_oracles,
         oracle_hint_service: OracleHintService::new(Some(versioned_program_path.as_std_path())),
     };